Added service layer in Booking microservice

+++ Microservices/booking-microservice/src/main/java/com/salonify/booking_microservice/model/SalonReport.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class SalonReport {
     private Long salonId;
     private String salonName;
-    private Double totalEarnings;
+    private int totalEarnings;
     private Integer totalBookings;
     private Integer cancelledBookings;
     private Double totalRefund;
